Extract some private methods from Header::ESP and put them in a mixin module Header::Crypto.
